Cursor fixed (hopefully finally). The X.org drivers moves a cursor although it just has been disabled (and it's BO removed therefore), which I find quite illegal, but allowing this fixes a jumpy cursor when it's hotspot changes. cursor_bo replaced by cursor_{offset,enabled,color} in lx_crtc.
Replaced wmb() by (the more generic? I hope so) DRM_WRITEMEMORYBARRIER().

Verified that it is indeed legal to wakeup userspace when we are signalled by
the vblank IRQ and a page-flip was queued. Renamed flip_pending to the more
accurate flip_scheduled, as it's determinable by reading DC_LINE_CNT whether a
flip is still pending or has already been latched by the hw (not checked as
flip_scheduled includes that (and more) information as well, since the code
is protected by locks).

Signed-off-by: Franz Brauße <dev@karlchenofhell.org>