several changes to command handling and copyrect().
Introduced explicit command buffer wrapping through a 8-byte sized dummy 'data'
command since the working of the HW is unnown to me and the spec isn't able to
shed some light there.
This fixes some issues encountered during testing.
Also it appears as if the command buffer and/or channel 3 need some time to
warm up. The amount of data passed through for the speed to catch up seems to
be directly related to the command buffer size (tried: 64 KB, 1 MB, 2 MB)...

The copyrect function now uses the GP's DMA channel 3 to copy source data from
the framebuffer. This tested somewhat and works at least for fbcon flawlessly.

Also the register generally used for starting BLTs and specifying it's source
only control the 'old' source DMA channel; if channel 3 is used, the BLT's
source must be programmed to 'no source required'. Thanks go out to AMD for
being so very precise in the data sheet; this took me two days to debug.

Additionally the GP's IRQs are enabled but not used now.

Signed-off-by: Franz Brauße <dev@karlchenofhell.org>